moved blade, hole, and trap classes into traps folder
