+++ src/main/java/org/bcit/com2522/project/traps/Trap.java
@@ -1,5 +1,7 @@
-package org.bcit.com2522.project;
+package org.bcit.com2522.project.traps;
 
+import org.bcit.com2522.project.Sprite;
+import org.bcit.com2522.project.Window;
 import processing.core.PVector;
 
 import java.awt.*;
